+++ lib/src/api/firebase_email_auth_api.dart
@@ -1,8 +1,13 @@
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'base_auth_api.dart';
 
 class FirebaseEmailAuthAPI implements BaseAuthAPI {
-  FirebaseEmailAuthAPI({this.email, this.password});
+  FirebaseEmailAuthAPI({
+    @required this.email,
+    @required this.password,
+  });
 
   final String email;
   final String password;
@@ -24,5 +29,10 @@ class FirebaseEmailAuthAPI implements BaseAuthAPI {
   @override
   Future<void> signOut() {
     return _firebaseAuth.signOut();
+  }
+
+  @override
+  Future<FirebaseUser> linkWith(FirebaseUser user) {
+    throw PlatformException(code: "UNSUPPORTED_FUNCTION_EXCEPTION", message: "e-mail sign-in does not support linking");
   }
 }