+++ CHANGELOG.md
@@ -1,7 +1,29 @@
+## [0.3.2] - 29/Jan/2020
+
+- Minor bug fix
+
+## [0.3.1] - 29/Jan/2020
+
+- Added getUserClaim function so that we can get claim data from provider
+- Changed logic of signing-out from 3rd party providers
+
+## [0.3.0] - 28/Jan/2020
+
+- Updated cloud function for verifiying/creating Kakao token
+- Changed logic of Kakao sign in
+- Refactoring Kakao sign in
+- Refactoring other sign ins
+- Added linking function
+- Made email and password as a required field
+
+## [0.2.0] - 27/Jan/2020
+
+- Supporting Facebook sign in
+
 ## [0.1.0] - 27/Jan/2020
 
 - Supporting Kakao sign in
 
 ## [0.0.1] - 25/Jan/2020
 
-- Initial release with email, Gmail support
+- Initial release with Firebase email and Gmail support